Validitation: Added Action to catch Persistence and Validation Exceptions
The CatchPersistenceAndValidationExceptionAction catches any Persistence
or Validation Exception which may occur during an update or creation of
an entry. The Exception is translated to an 409 Conflict Http Status
code. The Entries controller was altered accordingly. Further, two
dependencies were added to the build.sbt. Without them, the Validator
would not work.